+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - (Fixed issue #82) The incorrect mapping of mouse buttons under UNIX was fixed for **CUnixInputContext**.
 
+- The implementation of **CPhysics2DSystem::InjectBindings** was fixed.
+
 ### Added
 
 - The implementation of **CUnixInputContext::GetNormalizedMousePosition** was added.